moved the grasping using vlm pixel related implementation as a separate file in the skiulls/grasp_vlm.py file
