+++ exec_plan.py
@@ -8,7 +8,6 @@ plan provided as input.
 """
 
 import argparse
-import json
 from typing import Dict, Optional
 
 import numpy as np
@@ -19,6 +18,7 @@ from bosdyn.client.util import authenticate
 from numpy.typing import NDArray
 
 from skills.grasp import grasp_at_pixel
+from skills.grasp_vlm import grasp_with_vlm
 from skills.spot_hand_move import (
     close_gripper,
     move_hand_to_relative_pose,
@@ -29,7 +29,6 @@ from skills.wipe import wipe_multiple_strokes
 from skills.wipe_online import wipe_online as run_wipe_online
 from skills.push_button import push_button as run_push_button
 from skills.spot_navigation import navigate_to_absolute_pose
-from spot_utils.pretrained_model_interface import GoogleGeminiVLM
 from spot_utils.perception.spot_cameras import capture_images
 from spot_utils.spot_localization import SpotLocalizer
 from spot_utils.utils import (
@@ -38,9 +37,6 @@ from spot_utils.utils import (
     get_pixel_from_user,
     verify_estop,
 )
-import rerun as rr
-from PIL import Image
-import cv2
 
 DEFAULT_HAND_LOOK_FLOOR_POSE = math_helpers.SE3Pose(
     x=0.80, y=0.0, z=0.25, rot=math_helpers.Quat.from_pitch(np.pi / 3)
@@ -61,71 +57,6 @@ LOCALIZER = None
 ROBOT = None
 SAM_ENDPOINT = None
 SPOT_ROOM_POSE: Dict[str, float] = dict()
-
-
-def _get_pixel_from_gemini(vlm_query_str: str, pil_image: Image.Image) -> tuple[int, int]:
-    """Query Gemini VLM to get a single pixel [y, x] normalized to 0-1000, then
-    denormalize to image pixel coordinates.
-
-    This mirrors the usage pattern in the wipe tool: construct a
-    GoogleGeminiVLM instance and call sample_completions directly.
-    """
-
-    vlm = GoogleGeminiVLM("gemini-2.5-pro")
-
-    def _strip_markdown_fence(json_output_str: str) -> str:
-        """Remove ```json fences if present and return the inner JSON string."""
-        lines = json_output_str.splitlines()
-        for i, line in enumerate(lines):
-            if line.strip() == "```json":
-                json_output_str = "\n".join(lines[i + 1 :])
-                json_output_str = json_output_str.split("```")[0]
-                break
-        return json_output_str.strip()
-
-    # 1) Query the VLM
-    vlm_output_list = vlm.sample_completions(
-        prompt=vlm_query_str,
-        imgs=[pil_image],
-        temperature=0.0,
-        seed=42,
-        num_completions=1,
-    )
-    vlm_output_str = vlm_output_list[0]
-
-    # 2) Parse JSON output
-    json_string_to_parse = _strip_markdown_fence(vlm_output_str)
-    parsed_data = json.loads(json_string_to_parse)
-
-    if not isinstance(parsed_data, list) or not parsed_data:
-        raise ValueError("Parsed JSON is not a non-empty list.")
-
-    first_point_obj = parsed_data[0]
-    if (
-        "point" not in first_point_obj
-        or not isinstance(first_point_obj["point"], list)
-        or len(first_point_obj["point"]) != 2
-    ):
-        raise ValueError(
-            "First element in JSON does not contain a valid 'point' list [y, x]."
-        )
-
-    y_norm, x_norm = first_point_obj["point"]
-    if not isinstance(y_norm, (int, float)) or not isinstance(x_norm, (int, float)):
-        raise ValueError("Normalized coordinates are not numbers.")
-
-    # 3) Denormalize from 0–1000 range to image pixel coordinates
-    img_height = pil_image.height
-    img_width = pil_image.width
-    y = int(y_norm * img_height / 1000.0)
-    x = int(x_norm * img_width / 1000.0)
-
-    # Clamp to image bounds
-    y = max(0, min(y, img_height - 1))
-    x = max(0, min(x, img_width - 1))
-
-    # Return as (x, y) pixel coordinate
-    return (x, y)
 
 
 def np_pose_to_SE3(X_RobEE: NDArray) -> math_helpers.SE3Pose:
@@ -183,45 +114,9 @@ def gaze(direction: str) -> None:
 
 def grasp(text_prompt: Optional[str]) -> None:
     """Grasp an object at a specified pixel."""
-    # Capture an image.
-    camera = "hand_color_image"
     assert ROBOT is not None, "Sahit why!"
     assert LOCALIZER is not None, "SAHIT WHY!!!!"
-
-    images = capture_images(ROBOT, LOCALIZER, [camera])
-    rgbd = images[camera]
-    rgb_np = rgbd.rgb
-    rr.log("rgb_raw", rr.Image(rgb_np))
-
-    # FIXME: Don't do this!!! Hiding implementation
-    # if text_prompt and SAM_ENDPOINT:
-    #     # Select a pixel by querying GroundedSAM.
-    #     pixel = get_pixel_from_grounded_sam(rgbd.rgb, text_prompt, SAM_ENDPOINT)
-    # else:
-    #     # Select a pixel by querying the user.
-    #     pixel = get_pixel_from_user(rgbd.rgb)
-
-    # Call Gemini to point
-    vlm_query_template = f"""
-    Point to the {text_prompt}. If you cannot see the {text_prompt} fully, point to the best guess.
-    The answer should follow the json format: [{{"point": , "label": }}, ...]. The points are in [y, x] format normalized to 0-1000.
-    """
-    image_pil = Image.fromarray(rgb_np)
-    pixel = _get_pixel_from_gemini(vlm_query_template, image_pil)
-
-    # Draw pixel on the image
-    bgr = cv2.cvtColor(rgb_np, cv2.COLOR_RGB2BGR)
-    cv2.circle(bgr, pixel, 5, (0, 0, 255), -1)
-    rgb_annotated = cv2.cvtColor(bgr, cv2.COLOR_BGR2RGB)
-    rr.log("pointing", rr.Image(rgb_annotated))
-
-    if pixel is not None:
-        # Grasp at the pixel with a top-down grasp.
-        top_down_rot = math_helpers.Quat.from_pitch(np.pi / 2)
-        grasp_at_pixel(ROBOT, rgbd, pixel, grasp_rot=top_down_rot)
-        return
-    else:
-        raise RuntimeError("WTF. Grasp failed!")
+    grasp_with_vlm(ROBOT, LOCALIZER, text_prompt)
 
 
 def grasp_at_pose(X_RobEE: NDArray) -> None:
